Replaced GRM.GetLvlCap function with a global
The GRM code was calling the GRM.GetLvlCap function in many places, but the result of that function is always the same. So to optimize a bit, just define it as a global instead.

This utilizes the GetMaxLevelForPlayerExpansion() function if it exists, and GetMaxPlayerLevel() otherwise.

Also removed the GRM_G.SOD level cap workaround, since that should be handled by the global as well. (Retained the GRM_G.SOD global variable)